% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categories}
\alias{categories}
\title{Categories dataset}
\format{
A data frame with 1 column:
\describe{
\item{animal}{Factor with animal categories as letters}
}
}
\source{
Generated for examples
}
\usage{
categories
}
\description{
A simple dataset of categorical values used for examples.
}
\keyword{datasets}
