% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories2.R
\docType{data}
\name{categories2}
\alias{categories2}
\title{Categories2: Triangular Distribution of Bikinibottom Species}
\format{
A data frame with 250 rows and 11 variables:
\describe{
\item{animal}{Character. Species/animal names.}
\item{freq}{Integer. Frequency of each species, forming a triangular pattern.}
}
}
\usage{
categories2
}
\description{
A dataset of dummy nominal data inspired by characters/species
from the Bikini Bottom universe (SpongeBob SquarePants).
This dataset simulates a roughly triangular distribution of frequencies.
}
\examples{
ranked_barplot(categories2, "animal")
}
\keyword{datasets}
