% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars}
\alias{starwars}
\title{Star Wars dataset}
\format{
A data frame
}
\source{
dplyr::starwars
}
\usage{
starwars
}
\description{
Character info from Star Wars (from dplyr/ggplot2 examples)
}
\keyword{datasets}
